% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_tables.R
\name{tmod_chi_squared_test}
\alias{tmod_chi_squared_test}
\title{Apply Pearson's Chi-squared test with Yates' continuity correction}
\usage{
tmod_chi_squared_test(formula, data)
}
\arguments{
\item{formula}{an object of class "formula"}

\item{data}{a data frame object}
}
\value{
An itest object.
}
\description{
Apply Pearson's Chi-squared test with Yates' continuity correction
}
