% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tmod_contingency}
\alias{tmod_contingency}
\title{Produce a contingency table from a data and formula object.}
\usage{
tmod_contingency(formula, data)
}
\arguments{
\item{formula}{an object of class "formula"}

\item{data}{a data frame object}
}
\value{
A table object with named rows and columns
}
\description{
Produce a contingency table from a data and formula object.
}
