% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_tables.R
\name{tmod_fisher_test}
\alias{tmod_fisher_test}
\title{Apply Fisher's Exact Test for Count Data}
\usage{
tmod_fisher_test(formula, data)
}
\arguments{
\item{formula}{an object of class "formula"}

\item{data}{a data frame object}
}
\value{
An itest object.
}
\description{
Apply Fisher's Exact Test for Count Data
}
