% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{tmod_kendall_correlation_test}
\alias{tmod_kendall_correlation_test}
\title{Apply the Kendall tau correlation test}
\usage{
tmod_kendall_correlation_test(formula, data)
}
\arguments{
\item{formula}{an object of class "formula"}

\item{data}{a data frame object}
}
\value{
An itest object.
}
\description{
Apply the Kendall tau correlation test
}
