% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/means.R
\name{tmod_kruskal_wallis_test}
\alias{tmod_kruskal_wallis_test}
\title{Apply the Kruskal-Wallis test}
\usage{
tmod_kruskal_wallis_test(formula, data)
}
\arguments{
\item{formula}{an object of class "formula"}

\item{data}{a data frame object}
}
\value{
An itest object.
}
\description{
Apply the Kruskal-Wallis test
}
