% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{tmod_lin_reg_table}
\alias{tmod_lin_reg_table}
\title{Get a linear regression model table}
\usage{
tmod_lin_reg_table(formula, data)
}
\arguments{
\item{formula}{an object of class "formula"}

\item{data}{a data frame object}
}
\value{
An itest object.
}
\description{
Get a linear regression model table
}
