% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/means.R
\name{tmod_mann_whitney_test}
\alias{tmod_mann_whitney_test}
\title{Apply the two-sample Mann-Whitney Test}
\usage{
tmod_mann_whitney_test(formula, data)
}
\arguments{
\item{formula}{an object of class "formula"}

\item{data}{a data frame object}
}
\value{
An itest object.
}
\description{
Apply the two-sample Mann-Whitney Test
}
