% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tmod_mean_by_group}
\alias{tmod_mean_by_group}
\title{Compute group means.}
\usage{
tmod_mean_by_group(formula, data)
}
\arguments{
\item{formula}{an object of class "formula"}

\item{data}{a data frame object}
}
\value{
A named vector of means.
}
\description{
Compute group means.
}
