% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_tables.R
\name{tmod_odds_ratio}
\alias{tmod_odds_ratio}
\title{Apply Odds Ratio Test for Two or More Groups}
\usage{
tmod_odds_ratio(formula, data, effect = NULL, treatment = NULL)
}
\arguments{
\item{formula}{an object of class "formula"}

\item{data}{a data frame object}

\item{effect}{optional name of the 'effect' category in y}

\item{treatment}{optional name of the 'treatment' category in x}
}
\value{
An itest object.
}
\description{
Apply Odds Ratio Test for Two or More Groups
}
