% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/means.R
\name{tmod_one_way_anova_test}
\alias{tmod_one_way_anova_test}
\title{Apply the one-way Analysis of Variance test}
\usage{
tmod_one_way_anova_test(formula, data)
}
\arguments{
\item{formula}{an object of class "formula"}

\item{data}{a data frame object}
}
\value{
An itest object.
}
\description{
Apply the one-way Analysis of Variance test
}
