% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tmod_prop_by_group}
\alias{tmod_prop_by_group}
\title{Produce sample proportions.}
\usage{
tmod_prop_by_group(formula, data)
}
\arguments{
\item{formula}{an object of class "formula"}

\item{data}{a data frame object}
}
\value{
A matrix of proportions. Column totals should (approximately) sum to one.
}
\description{
Produce sample proportions.
}
