% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/means.R
\name{tmod_t_test}
\alias{tmod_t_test}
\title{Apply a two-sample t-test}
\usage{
tmod_t_test(formula, data)
}
\arguments{
\item{formula}{an object of class "formula"}

\item{data}{a data frame object}
}
\value{
An itest object.
}
\description{
Apply a two-sample t-test
}
