% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_sample.R
\name{tmod_t_test_one}
\alias{tmod_t_test_one}
\title{Apply One Sample T-Test}
\usage{
tmod_t_test_one(formula, data, h0 = 0)
}
\arguments{
\item{formula}{an object of class "formula"}

\item{data}{a data frame object}

\item{h0}{mean to test as the null-hypothesis}
}
\value{
An itest object.
}
\description{
Apply One Sample T-Test
}
