% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_sample.R
\name{tmod_wilcoxon_test}
\alias{tmod_wilcoxon_test}
\title{Apply the one-sample Wilcoxon Test}
\usage{
tmod_wilcoxon_test(formula, data, h0 = 0)
}
\arguments{
\item{formula}{an object of class "formula"}

\item{data}{a data frame object}

\item{h0}{mean to test as the null-hypothesis}
}
\value{
An itest object.
}
\description{
Apply the one-sample Wilcoxon Test
}
