% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_tables.R
\name{tmod_z_test_prop}
\alias{tmod_z_test_prop}
\title{Apply Z Test for Difference in Proportions Across Two Groups}
\usage{
tmod_z_test_prop(formula, data, effect = NULL, treatment = NULL)
}
\arguments{
\item{formula}{an object of class "formula"}

\item{data}{a data frame object}

\item{effect}{optional name of the 'effect' category in y}

\item{treatment}{optional name of the 'treatment' category in x}
}
\value{
An itest object.
}
\description{
Apply Z Test for Difference in Proportions Across Two Groups
}
