% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{tmod_logistic_regression}
\alias{tmod_logistic_regression}
\title{Apply a logistic regression model}
\usage{
tmod_logistic_regression(formula, data)
}
\arguments{
\item{formula}{an object of class "formula"}

\item{data}{a data frame object}
}
\value{
An itest object.
}
\description{
Apply a logistic regression model
}
